requirements.txt
alabaster==0.7.12
anyascii==0.1.7
appdirs==1.4.3
asgiref==3.5.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.9.0
#backports.csv==1.0.7
#backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==23.7.0
bleach==4.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.12.3
chardet==3.0.4
click==8.1.6
coderedcms==2.1.4
coderedcms-bootstrap4==2.0.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect==1.1.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.2.20
django-appconf==1.0.4
django-bootstrap4==22.2
django-bootstrap5==22.2
django-classy-tags==2.0.0
django-cookie-law==2.0.5
django-debug-toolbar==3.7.0
django-environ==0.4.5
django-eventtools==1.0.3
django-filter==2.4.0
django-honeypot==1.0.3
django-htmx==1.16.0
django-import-export==2.5.0
django-jazzmin==2.6.0
django-markdownify==0.9.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-sass==1.1.0
django-sendmail-backend==0.1.2
django-stubs==1.7.0
django-tables2==2.3.4
django-tables2-column-shifter==0.5.2
django-taggit==3.1.0
django-tempus-dominus==5.1.2.13
django-treebeard==4.5.1
django-waitress==0.1.0
django_tableaux==0.6.2
djangorestframework==3.12.2
djangorestframework-stubs==1.4.0
docutils==0.16
draftjs-exporter==2.1.6
et-xmlfile==1.0.1
Fabric3==1.14.post1
fake-useragent==0.1.11
Faker==0.9.1
future==0.17.1
geocoder==1.38.1
html5lib==1.0.1
icalendar==4.1.0
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
iniconfig==1.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
l18n==2020.6.1
libsass==0.22.0
lxml==4.9.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==7.2.0
mypy==0.800
mypy-extensions==0.4.3
odfpy==1.4.0
openpyxl==3.1.2
packaging==23.1
paramiko==2.6.0
parse==1.12.1
pathspec==0.11.1
Pillow==9.2.0
pipdeptree==2.13.0
platformdirs==3.9.1
pluggy==0.13.0
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.19
pyee==6.0.0
Pygments==2.7.4
PyNaCl==1.3.0
pyparsing==2.4.5
pyppeteer==0.0.25
pyquery==1.4.0
pytest==6.2.2
pytest-django==4.1.0
pytest-runner==5.3.0
python-dateutil==2.8.0
python-resize-image==1.1.19
pytz==2021.1
PyYAML==6.0
ratelim==0.1.6
rcssmin==1.0.6
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
rjsmin==1.1.0
sentry-sdk==0.11.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.1
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
tablib==3.0.0
telepath==0.2
text-unidecode==1.2
toml==0.10.2
tomli==2.0.1
tqdm==4.35.0
typed-ast==1.4.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.3
w3lib==1.21.0
wagtail==4.2.4
wagtail-cache==2.2.0
wagtail-seo==2.3.0
waitress==1.4.4
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.0.2
whitenoise==6.2.0
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==0.6.0
